+++ extensions/apps/extensions/src/components/pages/installed-extensions/index.tsx
@@ -1,9 +1,19 @@
 import React from 'react';
+import AppList from '@akashaorg/design-system-components/lib/components/AppList';
+import Button from '@akashaorg/design-system-core/lib/components/Button';
+import Card from '@akashaorg/design-system-core/lib/components/Card';
+import Image from '@akashaorg/design-system-core/lib/components/Image';
+import Stack from '@akashaorg/design-system-core/lib/components/Stack';
+import Text from '@akashaorg/design-system-core/lib/components/Text';
+import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
+import InfoCard from '@akashaorg/design-system-core/lib/components/InfoCard';
 import routes, { EXTENSIONS } from '../../../routes';
 import { useTranslation } from 'react-i18next';
+import { InstalledExtension, useInstalledExtensions } from './use-installed-extensions';
 import { useNavigate } from '@tanstack/react-router';
 import { useRootComponentProps } from '@akashaorg/ui-awf-hooks';
-import { InstalledExtensions } from './installed-extensions';
+
+const PUBLIC_IMAGE_PATH = '/images';
 
 export const InstalledExtensionsPage: React.FC<unknown> = () => {
   const navigate = useNavigate();
@@ -22,26 +32,93 @@ export const InstalledExtensionsPage: React.FC<unknown> = () => {
     });
   };
 
+  const { data, error, loading } = useInstalledExtensions();
+
+  const addAction = (ext: InstalledExtension) => ({
+    ...ext,
+    action: (
+      <Button variant="secondary" label={t('Open')} onClick={() => handleAppClick(ext.name)} />
+    ),
+  });
+
+  const installedExtensions = data?.map(addAction);
+  const defaultExtensions = [];
+
   return (
-    <InstalledExtensions
-      titleLabel={t('Installed Extensions')}
-      sections={[
-        {
-          assetName: 'longbeam-notfound',
-          title: t('No extensions installed yet!'),
-          discoverLabel: t('Discover'),
-          description: t('cool extensions and install them'),
-          description2: t('to customize your world'),
-          onClickDiscover: handleClickDiscover,
-        },
-        {
-          title: t('Default Extensions'),
-          description: t(
-            'The default extensions are the ones that come preinstalled with AKASHA World. You cannot uninstall them.',
-          ),
-        },
-      ]}
-      handleAppClick={handleAppClick}
-    />
+    <Stack spacing="gap-y-4">
+      <Text variant="h5">{t('Installed Extensions')}</Text>
+      <Card padding="p-4">
+        <Stack spacing="gap-y-3" align="center">
+          {loading && (
+            <Stack spacing="gap-y-5" align="center">
+              <Spinner />
+              <Text variant="button-md">{t('Loading installed extensions')}</Text>
+            </Stack>
+          )}
+          {error && (
+            <InfoCard
+              titleLabel={
+                <Text variant="h5" align="center">
+                  {t(`Uh-oh! We couldn't load`)} <br /> {t(`the extensions list!`)}
+                </Text>
+              }
+              bodyLabel={
+                <Text variant="body2" align="center" customStyle="w-60 sm:w-auto">
+                  {t('There might be an issue with the database. Please try again later!')}
+                </Text>
+              }
+              assetName="SearchApp_NotFound-min"
+            />
+          )}
+          {installedExtensions && (
+            <>
+              {!installedExtensions.length && (
+                <>
+                  <Stack customStyle="h-52 w-52">
+                    <Image
+                      customStyle="object-contain"
+                      src={`${PUBLIC_IMAGE_PATH}/longbeam-notfound.webp`}
+                    />
+                  </Stack>
+                  <Text variant="h6">{t('No extensions installed yet!')}</Text>
+                  <Stack align="center">
+                    <Text as="span" color={{ light: 'grey5', dark: 'grey6' }}>
+                      <Button
+                        variant="text"
+                        size="md"
+                        label={t('Discover')}
+                        onClick={handleClickDiscover}
+                      />{' '}
+                      {t('cool extensions and install them')}
+                    </Text>
+                    <Text variant="body2" color={{ light: 'grey5', dark: 'grey6' }}>
+                      {t('to customize your world')}
+                    </Text>
+                  </Stack>
+                </>
+              )}
+              {!!installedExtensions.length && (
+                <AppList
+                  apps={installedExtensions}
+                  //implementation requires pagination support on installed extensions service on sdk
+                  onLoadMore={() => null}
+                />
+              )}
+            </>
+          )}
+        </Stack>
+      </Card>
+      <Card padding="p-4" margin="mb-2">
+        <Stack spacing="gap-y-3">
+          <Text variant="h6">{t('Default Extensions')}</Text>
+          <Text variant="body2" color={{ light: 'grey5', dark: 'grey6' }}>
+            {t(
+              'The default extensions are the ones that come preinstalled with AKASHA World. You cannot uninstall them.',
+            )}
+          </Text>
+          <AppList apps={defaultExtensions} onLoadMore={() => null} />
+        </Stack>
+      </Card>
+    </Stack>
   );
 };
